requirements.txt
Django>=3.0.7,<3.1.0
djangorestframework>=3.11.0,<3.12.0
mysqlclient
djangorestframework-simplejwt
pyjwt